+++ creators/redis/Creator.go
@@ -1,4 +1,4 @@
-package mongo
+package redis
 
 import (
 	"fmt"
@@ -8,12 +8,12 @@ import (
 
 type Creator struct{}
 
-const port uint16 = 27017
+const port uint16 = 6379
 
 func (c Creator) GetDefaultOpts() dbCreator.DefaultOpts {
 	return dbCreator.DefaultOpts{
 		Port:      port,
-		User:      "mongo",
+		User:      "",
 		DockerTag: "latest",
 		Password:  "",
 	}
@@ -21,19 +21,17 @@ func (c Creator) GetDefaultOpts() dbCreator.DefaultOpts {
 
 func (c Creator) GetCapabilities() dbCreator.Capabilities {
 	return dbCreator.Capabilities{
-		DatabaseName: true,
-		UserPassword: true,
+		DatabaseName: false,
+		UserPassword: false,
 	}
 }
 
 func (c Creator) Create(shell dbCreator.Shell, opts dbCreator.CreateOptions) error {
-	// https://hub.docker.com/_/mongo
+	// https://hub.docker.com/_/redis/
 	return shell.Run("docker", "run", "--name", opts.ContainerName,
-		"-e", dbCreator.DockerEnv("MONGO_INITDB_ROOT_PASSWORD", opts.Password),
-		"-e", dbCreator.DockerEnv("MONGO_INITDB_ROOT_USERNAME", opts.User),
-		"-e", dbCreator.DockerEnv("MONGO_INITDB_DATABASE", opts.Database),
 		"-p", fmt.Sprintf("%d:%d", port, opts.Port),
-		"-d", fmt.Sprintf("mongo:%s", opts.DockerTag),
+		"-d", fmt.Sprintf("redis:%s", opts.DockerTag),
+		"redis-server", "--save", "60", "1", "--loglevel", "warning",
 	)
 }