+++ dbCreator/DbCreator.go
@@ -11,6 +11,13 @@ type CreateOptions struct {
 	DryRun        bool
 }
 
+// Creator capabilities list
+type Capabilities struct {
+	DatabaseName bool
+	UserPassword bool
+}
+
+// Creator default options
 type DefaultOpts struct {
 	User      string
 	DockerTag string
@@ -20,6 +27,7 @@ type DefaultOpts struct {
 
 type DbCreator interface {
 	GetDefaultOpts() DefaultOpts
+	GetCapabilities() Capabilities
 	Create(shell Shell, opts CreateOptions) error
 	ValidatePassword(password string) error
 }