+++ dbCreator/Shell.go
@@ -8,15 +8,18 @@ import (
 	"strings"
 )
 
+// Format DockerEnv variable for passing to docker options
 func DockerEnv(key string, value string) string {
 	return fmt.Sprintf("%s=%s", key, value)
 }
 
+// Child process runner with output verbosity flag
 type Shell struct {
 	dryRun  bool
 	verbose bool
 }
 
+// Create new shell instance
 func NewShell(dryRun bool, verbose bool) Shell {
 	return Shell{
 		dryRun:  dryRun,
@@ -24,6 +27,7 @@ func NewShell(dryRun bool, verbose bool) Shell {
 	}
 }
 
+// Create new shell instance
 func (sh Shell) RunWithOutput(name string, args ...string) (string, error) {
 	if sh.dryRun || sh.verbose {
 		fmt.Println(makeShellCmdString(name, args...))
@@ -36,10 +40,7 @@ func (sh Shell) RunWithOutput(name string, args ...string) (string, error) {
 	return string(out), err
 }
 
-/*
-Run a child process, printing its outputs to Stdout/Stderr only in the verbose
-mode
-*/
+// Run a child process, printing its outputs to Stdout/Stderr only in the verbose mode
 func (sh Shell) RunSilent(name string, args ...string) error {
 	if sh.dryRun || sh.verbose {
 		fmt.Println(makeShellCmdString(name, args...))
@@ -55,6 +56,7 @@ func (sh Shell) RunSilent(name string, args ...string) error {
 	return cmd.Run()
 }
 
+// Run a child process, printing its output to Stdout/Stderr
 func (sh Shell) Run(name string, args ...string) error {
 	if sh.dryRun || sh.verbose {
 		fmt.Println(makeShellCmdString(name, args...))
@@ -70,6 +72,7 @@ func (sh Shell) Run(name string, args ...string) error {
 
 var pattern *regexp.Regexp
 
+// Quote a shell argument
 func Quote(cmd string) string {
 	if pattern == nil {
 		pattern = regexp.MustCompile(`[^\w@%+=:,./-]`)