+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.1.0] - 2025.08.23
+
+### Added
+- Redis container creation
+
+### Fixed
+- Some typos in error messages
 
 ## [1.0.0] - 2025.04.02